Componentiza los elementos del chat.
